+++ Editor/Settings/AssemblyInfo.cs
@@ -1,4 +1,5 @@
 using System.Runtime.CompilerServices;
 
+[assembly: InternalsVisibleTo("ClarityConsole.Capture")]
 [assembly: InternalsVisibleTo("ClarityConsole.UI")]
 [assembly: InternalsVisibleTo("ClarityConsole.Tests")]